Move workflow mappings to common config
