+++ src/ggrc-client/js/models/mappers/mappings-ggrc.js
@@ -4,8 +4,11 @@
 */
 
 import {
+  Proxy,
   Direct,
   Search,
+  Multi,
+  CustomFilter,
 } from '../mappers/mapper-helpers';
 import Mappings from './mappings';
 import CustomAttributeDefinition from '../custom-attributes/custom-attribute-definition';
@@ -162,6 +165,65 @@ new Mappings('ggrc_core', {
   },
   AssessmentTemplate: {
     _related: ['Audit'],
+  },
+
+  // Workflow
+  TaskGroup: {
+    _canonical: [...coreObjects, 'Program'],
+  },
+  TaskGroupTask: {
+    _related: ['Person', 'Workflow'],
+  },
+  Workflow: {
+    _related: ['Person', 'TaskGroup', 'TaskGroupTask'],
+  },
+  CycleTaskGroupObjectTask: {
+    // It is needed for an object list generation. This object list
+    // describes which objects can be mapped to CycleTaskGroupObjectTask.
+    // Types placed within this collection will be intersected
+    // with TreeViewConfig.base_widgets_by_type["CycleTaskGroupObjectTask"]
+    // collection. The result of the operation is the total list.
+    _canonical: [...coreObjects, 'Audit', 'Program'],
+    _related: ['Person'],
+    // Needed for related_objects mapper
+    related_objects_as_source: Proxy(
+      null,
+      'destination', 'Relationship',
+      'source', 'related_destinations'
+    ),
+    // Needed for related_objects mapper
+    related_objects_as_destination: Proxy(
+      null,
+      'source', 'Relationship',
+      'destination', 'related_sources'
+    ),
+    // Needed to show mapped objects for CycleTaskGroupObjectTask
+    related_objects: Multi(
+      ['related_objects_as_source', 'related_objects_as_destination']
+    ),
+    /**
+       * "cycle" mapper is needed for mapped objects under
+       * CycleTaskGroupObjectTask into mapping-tree-view component.
+       */
+    cycle: Direct(
+      'Cycle', 'cycle_task_group_object_tasks', 'cycle'),
+    /**
+       * This mapping name is needed for objects mapped to CTGOT.
+       * It helps to filter results of objects mapped to CTGOT.
+       * We can just remove some objects from results.
+       */
+    info_related_objects: CustomFilter(
+      'related_objects',
+      function (relatedObjects) {
+        return !_.includes([
+          'CycleTaskGroup',
+          'Comment',
+          'Document',
+          'Person',
+        ],
+        relatedObjects.instance.type);
+      }
+    ),
   },
 
   // Other
