--- src/ggrc-client/js/models/mappers/mappings-workflow.js
@@ -1,83 +1,0 @@
-/*
-    Copyright (C) 2019 Google Inc.
-    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-*/
-
-import {
-  Proxy,
-  Direct,
-  Multi,
-  CustomFilter,
-} from './mapper-helpers';
-import Mappings from './mappings';
-
-const _workflowObjectTypes = [
-  'Program', 'Regulation', 'Policy', 'Standard', 'Contract',
-  'Requirement', 'Control', 'Objective', 'OrgGroup', 'Vendor', 'AccessGroup',
-  'System', 'Process', 'DataAsset', 'Product', 'Project', 'Facility',
-  'Market', 'Issue', 'Risk', 'Threat', 'Metric', 'TechnologyEnvironment',
-  'ProductGroup'];
-
-// Configure mapping extensions for ggrc_workflows
-// Add mappings for basic workflow objects
-let mappings = {
-  TaskGroup: {
-    _canonical: _workflowObjectTypes,
-  },
-  TaskGroupTask: {
-    _related: ['Person', 'Workflow'],
-  },
-  Workflow: {
-    _related: ['Person', 'TaskGroup', 'TaskGroupTask'],
-  },
-  CycleTaskGroupObjectTask: {
-    // It is needed for an object list generation. This object list
-    // describes which objects can be mapped to CycleTaskGroupObjectTask.
-    // Types placed within this collection will be intersected
-    // with TreeViewConfig.base_widgets_by_type["CycleTaskGroupObjectTask"]
-    // collection. The result of the operation is the total list.
-    _canonical: _workflowObjectTypes.concat('Audit'),
-    _related: ['Person'],
-    // Needed for related_objects mapper
-    related_objects_as_source: Proxy(
-      null,
-      'destination', 'Relationship',
-      'source', 'related_destinations'
-    ),
-    // Needed for related_objects mapper
-    related_objects_as_destination: Proxy(
-      null,
-      'source', 'Relationship',
-      'destination', 'related_sources'
-    ),
-    // Needed to show mapped objects for CycleTaskGroupObjectTask
-    related_objects: Multi(
-      ['related_objects_as_source', 'related_objects_as_destination']
-    ),
-    /**
-     * "cycle" mapper is needed for mapped objects under
-     * CycleTaskGroupObjectTask into mapping-tree-view component.
-     */
-    cycle: Direct(
-      'Cycle', 'cycle_task_group_object_tasks', 'cycle'),
-    /**
-     * This mapping name is needed for objects mapped to CTGOT.
-     * It helps to filter results of objects mapped to CTGOT.
-     * We can just remove some objects from results.
-     */
-    info_related_objects: CustomFilter(
-      'related_objects',
-      function (relatedObjects) {
-        return !_.includes([
-          'CycleTaskGroup',
-          'Comment',
-          'Document',
-          'Person',
-        ],
-        relatedObjects.instance.type);
-      }
-    ),
-  },
-};
-
-new Mappings('ggrc_workflows', mappings);
